requirements.txt
pillow==10.4.0
sqlparse==0.5.0
asgiref==3.8.1
django-markdownx==4.0.7
crispy-bootstrap5==0.6
Markdown==3.7
django-crispy-forms==1.13.0
typing_extensions==4.12.0
Django==4.0.10
